Move datastore package within data package. Also JwtInterceptor.kt should now get the token from the datastore
