+++ app/src/main/java/com/douglasbruce/tasky/core/data/datastore/TaskyPreferencesDataSource.kt
@@ -1,4 +1,4 @@
-package com.douglasbruce.tasky.core.datastore
+package com.douglasbruce.tasky.core.data.datastore
 
 import androidx.datastore.core.DataStore
 import com.douglasbruce.tasky.UserPreferences
