+++ app/src/main/java/com/douglasbruce/tasky/core/data/repository/OfflineFirstUserDataRepository.kt
@@ -1,6 +1,6 @@
 package com.douglasbruce.tasky.core.data.repository
 
-import com.douglasbruce.tasky.core.datastore.TaskyPreferencesDataSource
+import com.douglasbruce.tasky.core.data.datastore.TaskyPreferencesDataSource
 import com.douglasbruce.tasky.core.domain.repository.UserDataRepository
 import com.douglasbruce.tasky.core.model.UserData
 import kotlinx.coroutines.flow.Flow
